Convert the TitledCard into typescript + Create Props type
